Fix crash when decoding commercial certificates
The Erlang ssl application assumes that it is fully aware of all existing extensions, that they are in its list of SSL extensions, and it is able to decode them.

However since this is not true commercial certificates containing the id-pe-logotype extension crash the decoder. This is only one example of an extension that the ssl application is unaware of.

This patch improves the behaviour by returning the OID tuple and raw extension data for extensions that cannot be decoded.